+++ packages/rest-client/src/main/index.ts
@@ -5,7 +5,6 @@ export { default as asSystem } from './helpers/asSystem'
 export type { ApiConfig } from './types/ApiConfig'
 export { AgentConfig } from './types/ApiConfig'
 export type { AgentOptions } from './types/ApiConfig'
-export type { TransportConfig } from './types/ApiConfig'
 
 export type { AuthOptions } from './types/AuthOptions'